Add files via upload
All of the project files are uploaded here that was used for the final paper. This project exercises how front-end and back-end development work with database management systems using SQL, PHP, HTML, etc., and in the report, we successfully had the website accept new information to be stored.

Each of the files details are explained in the "Final Project Report.pdf" as well.